Check event queue on null close
When calling _tcp_close_api, guarantee that no events for that client
remain in the queue, even if the pcb was shut down.  This fixes a race
where an error callback is processed by the LwIP thread, invalidating
the pcb*, after the async thread has dequeued an event that will lead
to a close (such as a fin or final ack) but before close() is called.